clean up some things
remove redundant g argument
