+++ docs/Config.md
@@ -28,6 +28,14 @@ gui:
       - blue
   returnImmediately: false
   wrapMainPanel: true
+  # Side panel width as a ratio of the screen's width
+  sidePanelWidth: 0.333
+  # Determines whether we show the bottom line (the one containing keybinding
+	# info and the status of the app).
+  showBottomLine: true
+  # When true, increases vertical space used by focused side panel,
+  # creating an accordion effect
+  expandFocusedSidePanel: false
 logs:
   timestamps: false
   since: '60m'
